@@ -12,7 +12,7 @@ plugins {
 group 'org.hansi_b'
 version gitVersion()
 
-sourceCompatibility = 1.8
+sourceCompatibility = 11
 
 repositories {
 	// Use jcenter for resolving your dependencies.
@@ -22,14 +22,21 @@ repositories {
 
 dependencies {
 
-	testImplementation 'junit:junit:4.13'
+	//testImplementation 'junit:junit:4.13'
 	testImplementation 'org.assertj:assertj-core:3.17.1'
 	testImplementation 'org.codehaus.groovy:groovy:3.0.5'
+
+	testImplementation 'org.junit.vintage:junit-vintage-engine:5.6.2'
+	testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.6.2'
 
 	testImplementation 'org.spockframework:spock-core:1.3-groovy-3.0.5'
 
 	testImplementation 'org.objenesis:objenesis:3.1'
 	testImplementation 'org.spockframework:spock-core:2.0-M3-groovy-3.0'
+}
+
+test {
+	useJUnitPlatform()
 }
 
 jacocoTestReport {
