@@ -2,11 +2,11 @@ package org.hansi_b.moss.explain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hansi_b.moss.explain.MoveAsserts.*;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.hansi_b.moss.Sudoku;
 import org.hansi_b.moss.explain.Move.Strategy;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class TrivialNakedSingleTest {
 
